[Decomp][Spell] Cohesion split: extract periodic/stat aura handlers to SpellAuraPeriodic.cpp
Move the periodic/regen/resistance/stat/combat-rating aura block
(HandleAuraProcTriggerSpell through HandleNoReagentUseAura) out of
SpellAuras.cpp into a new SpellAuraPeriodic.cpp translation unit.

Pure file-level cohesion split: declarations stay in SpellAuras.h, the
AuraHandler[] table stays in SpellAuras.cpp and resolves the moved member
methods by address at link time. Byte-identical moves, no behaviour change.
SpellAuras.cpp 9040 -> 6151.
